requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
webdavclient3==3.14.6
langchain==0.0.340
openai==1.3.5
anthropic==0.5.0
python-multipart==0.0.6